Disable the contact detail photo popup menu (part 2)
This disables the menu altogether (the photo blinks when tapped but there
is no fading animation and no menu pops up), at least in combination with
part 1. It's possible that part 1 is unnecessary, but I'd rather err on
the side of having a comprehensive hack.

I manually backported it from cm-10.1-rc2-bkn (commit
609eebf4ac7a961ec6b6bd5a91ba1d40815b4e77).